LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/url-whitespace-consumption-expected.txt


PASS whitespace is optional between url( token and the string token
